Add dependency to qpid_proton gem
The Gem is required by Nuage network provider. However, because of a bug
in the published version of qpid_proton gem (0.17), the gem cannot be
used to succesfully authenticate with Nuage ActiveMQ. Because the bug
has been resolved, we temporarily propose addition of a custom Gem
version with that patch applied so that changes to Nuage provider can be
merged.

This dependency will be removed as soon as 0.18 release of qpid_proton
gem is available.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>